Refactor FTs so each test can be run in isolation
